Fixes to virtual methods
Fixed:
Only a single call could be made to each virtual method because the
definition was being mangled in the scope. Fixed with nVTAB_ACCESS.
Virutal methods couldn't be called with complex arguments, e.g.
obj.vmeth(com.plex). As the scratch reg used to access the vtab would be
overwritten. Fixed in IndexCall by making sure that args are pushed
before the function is located.